Modularizing the code
